more renaming and re-organization
