+++ src/data/index.ts
@@ -1,0 +1,3 @@
+import chunkData from './chunk-data';
+
+export default chunkData;
