+++ src/models/chunk.ts
