+++ src/models/clue.ts
