+++ src/models/index.ts
@@ -1,3 +1,3 @@
-export type { default as Chunk } from './Chunk';
-export type { default as Clue, ClueDifficulty } from './Clue';
-export { default as MapChunk } from './MapChunk';
+export type { default as Chunk } from './chunk';
+export type { default as Clue, ClueDifficulty } from './clue';
+export { default as MapChunk } from './map-chunk';
