+++ src/models/map-chunk.ts
